Admin Type , Update 미구현

+++ mysite03/src/main/java/com/douzone/mysite/repository/BoardRepository.java
@@ -50,7 +50,7 @@ public class BoardRepository {
 	}
 	
 	
-	public int findAllCount() {
+	public double findAllCount() {
 		
 		System.out.println("findAllCount()");
 		return sqlSession.selectOne("board.findAllCount");
